refact(ProductReviews): move dialog to separate component
